+++ pinery-miso/src/main/java/ca/on/oicr/pinery/lims/miso/MisoClient.java
@@ -967,7 +967,8 @@ public class MisoClient implements Lims {
         }
       }, //
       SEQUENCING_CONTROL_TYPE("sequencing_control_type", "Sequencing Control Type"), //
-      CUSTODY("custody", "Custody");
+      CUSTODY("custody", "Custody"), //
+      LATEST_TRANSFER_REQUEST("latest_transfer_request", "Latest Transfer Request");
 
       private final String sqlKey;
       private final String attributeKey;